Shows how to use JBoss Marshalling to serialize and deserialize an existing object model.
Shows how to serialize and deserialize an existing object model with JBoss Marshalling
without refactoring the model to fit into the serialization framework, and still get
an excellent performance.

The example also shows how to write an ObjectResolver to intercept objects
in the serialization and deserialization process. For this example I have created an
object resolver to detach and unproxy hibernate objects